Grid items styling and dropdown options
