+++ manifest-editor/src/components/layout/ThumbnailStripView.tsx
@@ -1,8 +1,9 @@
-import { ThumbnailStrip } from "../organisms/ThumbnailStrip";
 import { Tree } from "../../navigation/tree/Tree";
 
 import styled from "styled-components";
 import { ViewSelector } from "../../atoms/ViewSelector";
+import { GridList } from "../organisms/GridView/GridList";
+import { FlexContainerColumn } from "./FlexContainer";
 
 const ContentSelectorContainer = styled.div`
    {
@@ -11,9 +12,10 @@ const ContentSelectorContainer = styled.div`
     border-right: 0.016rem solid #dddddd;
     justify-content: space-between;
     align-items: center;
-    margin: 0.375rem 0;
+    margin: 0.375rem;
+    overflow-y: scroll;
     padding: ${(props: any) => props.theme.padding.small || "0.5rem"} 0;
-    height: 80vh;
+    height: 75vh;
     @media (max-width: ${(props: any) => props.theme.device.tablet || "770px"}) {
       width: 100%;
       border-right: none;
@@ -28,11 +30,13 @@ export const ThumbnailStripView: React.FC<{
   return (
     <>
       {(view === "thumbnails" || view === "tree") && (
-        <ContentSelectorContainer>
-          {view === "thumbnails" && <ThumbnailStrip />}
-          {view === "tree" && <Tree />}
+        <FlexContainerColumn style={{ alignItems: "center" }}>
+          <ContentSelectorContainer>
+            {view === "thumbnails" && <GridList />}
+            {view === "tree" && <Tree />}
+          </ContentSelectorContainer>
           <ViewSelector />
-        </ContentSelectorContainer>
+        </FlexContainerColumn>
       )}
     </>
   );
